authentcation and authorization

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import get_object_or_404, render, redirect
+
 from .models import Product
 from .forms import AddProductForm
 
@@ -21,38 +22,47 @@ def product_details(request, pk):
 
 
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
+            form = AddProductForm()
+        context = {
+            "form": form,
+        }
+        return render(request, 'products/product-add.html', context)
     else:
-        form = AddProductForm()
-    context = {
-        "form": form,
-    }
-    return render(request, 'products/product-add.html', context)
+        return redirect('products_list')
 
 
 def product_edit(request, pk):
-    product = get_object_or_404(Product, pk=pk)
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
+        if request.method == 'POST':
+            form = AddProductForm(
+                request.POST, request.FILES, instance=product)
 
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES, instance=product)
-
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
+            form = AddProductForm(instance=product)
+        context = {
+            "form": form,
+        }
+        return render(request, 'products/product-add.html', context)
     else:
-        form = AddProductForm(instance=product)
-    context = {
-        "form": form,
-    }
-    return render(request, 'products/product-add.html', context)
+        return redirect('products_list')
 
 
 def product_delete(request, pk):
-    product = get_object_or_404(Product, pk=pk)
-    product.delete()
-    return redirect('products_list')
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
+        product.delete()
+        return redirect('products_list')
+    else:
+        return redirect('products_list')
